AMQP-275 Race Conditions When Stopping Container
It was possible for messages to be left un-acked because they
were redelivered to the queue before the cancel.

It was possible for some messages to be un-acked between issuing
the cancel and receiving the cancelOk.

Even with both these cases resolved, we still observed occasional
un-acked messages.

Physically closing the channel (rather than returning it to the
cache) when stopping the container resolves this issue.

Add code to send the cancel immediately when stopping the container.

Add code to wait for the cancelOk and wait for the consumer threads
to stop separately.

Remove code that discarded deliveries between cancel and cancelOk.

Add code to force close the channel on container stop.

Add test case.