Workaround for Exchange/RK Reply Routing
If you use java config (instead of XML)  to configure the template, you
can "fake out" the reply queue name in the template so
it actually contains the `replyTo` address with the form `exchange/routingKey`.

This can't be done with the XML namespace because the same attribute is used
in both places.