AMQP-318: PoC Only
Lots of TODOs; I am thinking this is just not worth the effort.

The basic premise is, if we failed over to a secondard server, we should
try and fail back to the primary server every so often.

However, we must let existing channels/connections to quiesce and only
close them when the user returns them to the cache.